# packages/schemas/openapi.json
{
  "components": {
    "schemas": {
      "AgentDef": {
        "description": "One row of `agent_defs` (\u00a74). Frozen: a run holds a snapshot of it.",
        "properties": {
          "allowed_tools": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Allowed Tools",
            "type": "array"
          },
          "auto_approve": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RiskLevel"
            },
            "title": "Auto Approve",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_builtin": {
            "default": false,
            "title": "Is Builtin",
            "type": "boolean"
          },
          "max_steps": {
            "default": 20,
            "minimum": 1.0,
            "title": "Max Steps",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "space_id": {
            "title": "Space Id",
            "type": "string"
          },
          "system_prompt": {
            "title": "System Prompt",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "name",
          "role",
          "system_prompt",
          "created_at",
          "updated_at"
        ],
        "title": "AgentDef",
        "type": "object"
      },
...
      "CopyFrom": {
        "description": "Seed a new space with copies of another space's roster.",
        "properties": {
          "copy_from": {
            "title": "Copy From",
            "type": "string"
          }
        },
        "required": [
          "copy_from"
        ],
        "title": "CopyFrom",
        "type": "object"
      },
...
      "CreateSpaceRequest": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "seed": {
            "anyOf": [
              {
                "enum": [
                  "empty",
                  "builtins"
                ],
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/CopyFrom"
              }
            ],
            "default": "builtins",
            "title": "Seed"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateSpaceRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RiskLevel": {
        "description": "How much damage a tool call can do. Policy is a *set* of levels, never a threshold.",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "SpaceResponse": {
        "description": "A space, plus where its runs read and write.",
        "properties": {
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "auto_approve": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RiskLevel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Approve"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "max_agents_per_run": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Agents Per Run"
          },
          "max_run_cost_micros": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Run Cost Micros"
          },
          "max_run_seconds": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Run Seconds"
          },
          "max_steps_per_agent": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps Per Agent"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "tool_policies": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/ToolPolicy"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Policies"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at",
          "folder",
          "is_default"
        ],
        "title": "SpaceResponse",
        "type": "object"
      },
...
      "ToolPolicy": {
        "description": "What the gate does with a call to one tool, by name, before its risk level is read.\n\n``ASK`` is the default and means the risk-level rule decides; ``ALLOW``\nruns the call without asking; ``DENY`` refuses it without asking. A tool\nabsent from a policy is ``ASK``.",
        "enum": [
          "ask",
          "allow",
          "deny"
        ],
        "title": "ToolPolicy",
        "type": "string"
      },
...
      "UpdateSpaceRequest": {
        "additionalProperties": false,
        "description": "A partial update. ``None`` means \"inherit\", so `model_fields_set` tells it from \"not sent\".",
        "properties": {
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          },
          "auto_approve": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RiskLevel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Approve"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "max_agents_per_run": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Agents Per Run"
          },
          "max_run_cost_micros": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Run Cost Micros"
          },
          "max_run_seconds": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Run Seconds"
          },
          "max_steps_per_agent": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps Per Agent"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 60,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "tool_policies": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/ToolPolicy"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Policies"
          }
        },
        "title": "UpdateSpaceRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AgentBase sidecar",
    "version": "0.4.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/spaces": {
      "get": {
        "description": "Every space, archived ones included: their runs are still viewable.",
        "operationId": "list_spaces_spaces_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SpaceResponse"
                  },
                  "title": "Response List Spaces Spaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Spaces"
      },
      "post": {
        "description": "Create a space and seed its roster.\n\nThe seed is its own write after the row exists, so a partial copy leaves\na space the user can see and fix.",
        "operationId": "create_space_spaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Space"
      }
    },
    "/spaces/{space_id}": {
      "delete": {
        "description": "Delete an empty space and its agents. 409 for the default, or one with runs.",
        "operationId": "delete_space_spaces__space_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Space"
      },
      "get": {
        "operationId": "get_space_spaces__space_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Space"
      },
      "patch": {
        "operationId": "update_space_spaces__space_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Space"
      }
    },
...
    "/spaces/{space_id}/seed": {
      "post": {
        "description": "Add the starter roles an existing roster lacks. Safe to press twice.",
        "operationId": "seed_space_spaces__space_id__seed_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentDef"
                  },
                  "title": "Response Seed Space Spaces  Space Id  Seed Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Seed Space"
      }
    },
...
  }
}